safe_path_finder/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
blinker==1.8.2
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
distro==1.9.0
et-xmlfile==2.0.0
exceptiongroup==1.3.1
flask==3.0.3
groq==0.33.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.2.43
langchain-groq==0.1.10
langsmith==0.1.147
MarkupSafe==2.1.5
numpy==1.24.4
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.0.3
pydantic==2.10.6
pydantic-core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.4
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
tenacity==8.5.0
typing-extensions==4.13.2
tzdata==2026.1
urllib3==2.2.3
werkzeug==3.0.6
zipp==3.20.2
